feat: type 분리
