+++ frontend/src/types/requestType.ts
@@ -1,0 +1,10 @@
+export type InterceptHandler = (response: Response) => void;
+
+export type Interceptor = {
+  onSuccess: InterceptHandler;
+  onError: InterceptHandler;
+  setInterceptor: (
+    onSuccess?: InterceptHandler,
+    onError?: InterceptHandler
+  ) => void;
+};
